+++ Wab.Core/Service/DailyPoint/IPointCalculator.cs
@@ -1,8 +1,6 @@
-using System.Numerics;
-
 namespace Wab.Core.Service.DailyPoint;
 
 public interface IPointCalculator
 {
-    public BigInteger Calculate(DateTime cardObtainedDate);
+    public long Calculate(DateTime cardObtainedDate);
 }
